+++ fabric/src/main/java/top/theillusivec4/polymorph/mixin/core/AccessorHandledScreen.java
@@ -30,4 +30,7 @@ public interface AccessorHandledScreen {
 
   @Accessor
   int getY();
+
+  @Accessor
+  int getBackgroundHeight();
 }